Add files via upload
The project focus mainly about preventing the message from being hijacked or retransformed by an third party users. TCL language is been used in this project for better understanding of te activity that takes place while a data is being transferred. Watchdog timer is been implemented inorder to detect the attacks and prevent it.